feat(editor): adjustments to tests to work with standalone components
