+++ apps/metadata-editor/src/app/dashboard/search-header/search-header.component.ts
@@ -5,8 +5,8 @@ import {
   AuthService,
   AvatarServiceInterface,
 } from '@geonetwork-ui/feature/auth'
+import { LetDirective } from '@ngrx/component'
 import { FeatureSearchModule } from '@geonetwork-ui/feature/search'
-import { LetModule } from '@ngrx/component'
 
 @Component({
   selector: 'md-editor-search-header',
@@ -14,7 +14,7 @@ import { LetModule } from '@ngrx/component'
   styleUrls: ['./search-header.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
-  imports: [FeatureSearchModule, MatIconModule, CommonModule, LetModule],
+  imports: [FeatureSearchModule, MatIconModule, CommonModule, LetDirective],
 })
 export class SearchHeaderComponent {
   public placeholder = this.avatarService.placeholder
