+++ apps/metadata-editor/src/app/dashboard/sidebar/sidebar.component.spec.ts
@@ -1,43 +1,9 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  NO_ERRORS_SCHEMA,
-} from '@angular/core'
+import { ChangeDetectionStrategy, NO_ERRORS_SCHEMA } from '@angular/core'
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-import { AuthService } from '@geonetwork-ui/feature/auth'
-import { USER_FIXTURE } from '@geonetwork-ui/common/fixtures'
-import { LetDirective } from '@ngrx/component'
-import { BehaviorSubject, of } from 'rxjs'
+import { of } from 'rxjs'
 import { SidebarComponent } from './sidebar.component'
-import { UserModel } from '@geonetwork-ui/common/domain/user.model'
 import { ActivatedRoute } from '@angular/router'
-import { TranslateService } from '@ngx-translate/core'
-
-@Component({
-  // eslint-disable-next-line
-  selector: 'md-editor-dashboard-menu',
-  template: '<div></div>',
-})
-class DashboardMenuMockComponent {}
-
-@Component({
-  // eslint-disable-next-line
-  selector: 'gn-ui-user-preview',
-  template: '<div></div>',
-})
-export class UserPreviewComponent {
-  @Input() user: UserModel
-}
-
-const user = USER_FIXTURE()
-class AuthServiceMock {
-  user$ = new BehaviorSubject(user)
-}
-
-const translateServiceMock = {
-  currentLang: 'fr',
-}
+import { TranslateModule } from '@ngx-translate/core'
 
 describe('SidebarComponent', () => {
   let component: SidebarComponent
@@ -45,16 +11,11 @@ describe('SidebarComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [SidebarComponent],
+      imports: [SidebarComponent, TranslateModule.forRoot()],
       providers: [
-        { provide: AuthService, useClass: AuthServiceMock },
         {
           provide: ActivatedRoute,
           useValue: { params: of({ id: 1 }) },
-        },
-        {
-          provide: TranslateService,
-          useValue: translateServiceMock,
         },
       ],
       schemas: [NO_ERRORS_SCHEMA],
