+++ apps/metadata-editor/src/app/records/all-records/all-records-list.component.ts
@@ -4,8 +4,8 @@ import { MatIconModule } from '@angular/material/icon'
 import { Router } from '@angular/router'
 import { CatalogRecord } from '@geonetwork-ui/common/domain/record'
 import { SearchFacade } from '@geonetwork-ui/feature/search'
-import { UiSearchModule } from '@geonetwork-ui/ui/search'
 import { DashboardSearchService } from '../../dashboard/dashboard-search.service'
+import { UiSearchModule } from '@geonetwork-ui/ui/search'
 
 @Component({
   selector: 'md-editor-all-records-list',
