+++ libs/ui/elements/src/index.ts
@@ -5,3 +5,4 @@ export * from './lib/metadata-catalog/metadata-catalog.component'
 export * from './lib/search-results-error/search-results-error.component'
 export * from './lib/thumbnail/thumbnail.component'
 export * from './lib/content-ghost/content-ghost.component'
+export * from './lib/pagination-buttons/pagination-buttons.component'
