+++ libs/ui/search/src/index.ts
@@ -4,6 +4,6 @@ export * from './lib/facets/fixtures'
 export * from './lib/ui-search.module'
 export * from './lib/results-list/results-layout.config'
 export * from './lib/record-preview/record-preview.component'
-
 export * from './lib/record-preview-feed/record-preview-feed.component'
 export * from './lib/record-preview-row/record-preview-row.component'
+export * from './lib/record-table/record-table.component'
